+++ NEWS.md
@@ -33,3 +33,9 @@
 
 *Added version check to `run_linear_models()` so the user knows if an update is needed or not.
 *Added option to not include p-values, which will switch to using lme4 to run mixed-effect models.
+
+# turtle 0.1.3
+
+## BugFix
+
+*Fixed bug so that tidy output will optionally select p-values based on whether it exists in the output.